Moves CrankNicholson source and header files to /src and /include
