+++ SConstruct
@@ -81,7 +81,7 @@ src_files = Split("""src/TEM.cpp
                      src/snowsoil/SoilParent_Env.cpp
                      src/snowsoil/Stefan.cpp
                      src/snowsoil/TemperatureUpdator.cpp
-                     src/util/CrankNicholson.cpp
+                     src/CrankNicholson.cpp
                      src/util/tbc-debug-util.cpp
                      src/vegetation/Vegetation_Bgc.cpp
                      src/vegetation/Vegetation_Env.cpp""")
